Move maintenance commands to correct place
