+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/MaintenanceCommands.cs
@@ -12,6 +12,8 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands
     [Subcommand(typeof(DeleteNonPreservedScoresCommand))]
     [Subcommand(typeof(MarkNonPreservedScoresCommand))]
     [Subcommand(typeof(MigratePlaylistScoresToSoloScoresCommand))]
+    [Subcommand(typeof(VerifyImportedScoresCommand))]
+    [Subcommand(typeof(DeleteImportedHighScoresCommand))]
     public sealed class MaintenanceCommands
     {
         public Task<int> OnExecuteAsync(CommandLineApplication app, CancellationToken cancellationToken)
