+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/QueueCommands.cs
@@ -15,8 +15,6 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands
     [Subcommand(typeof(WatchQueueCommand))]
     [Subcommand(typeof(ImportHighScoresCommand))]
     [Subcommand(typeof(WatchHighScoresCommand))]
-    [Subcommand(typeof(VerifyImportedScoresCommand))]
-    [Subcommand(typeof(DeleteImportedHighScoresCommand))]
     public sealed class QueueCommands
     {
         public Task<int> OnExecuteAsync(CommandLineApplication app, CancellationToken cancellationToken)
